requirements.txt
flask
flask-cors
qrcode
pillow
cloudinary
python-dotenv
gunicorn